+++ src/matcher.rs
@@ -92,13 +92,13 @@ impl MultiFilePieceMatcher {
 
     fn read_bytes(
         path: &PathBuf,
-        read_length: usize,
-        read_start_position: usize,
+        read_length: u64,
+        read_start_position: u64,
     ) -> Result<Vec<u8>, std::io::Error> {
-        let mut read_bytes = vec![0u8; read_length];
+        let mut read_bytes = vec![0u8; read_length as usize];
         let mut handle = File::open(path)?;
 
-        handle.seek(SeekFrom::Start(read_start_position as u64))?;
+        handle.seek(SeekFrom::Start(read_start_position))?;
         handle.read_exact(&mut read_bytes)?;
 
         Ok(read_bytes)
